# docs/swagger/swagger.yaml
basePath: /
definitions:
  handlers.AddChannelRequest:
    properties:
      name:
        description: 頻道名稱
        example: 我的頻道
        type: string
      tags:
        description: 標籤列表
        items:
          type: integer
        type: array
    required:
    - name
    type: object
...
  handlers.SaveChannelRequest:
    properties:
      desc:
        type: string
      id:
        type: string
      name:
        type: string
      tags:
        items:
          type: integer
        type: array
    required:
    - id
    - name
    type: object
...
  handlers.SetChannelOwnerRequest:
    properties:
      c:
        description: 頻道 ID（歷史遺留參數，與 id 相同）
        type: string
      email:
        description: 要新增為共用者的使用者 Email（優先使用）
        type: string
      id:
        description: 頻道 ID
        type: string
      owners:
        description: 要新增的使用者 ID 陣列（向後相容）
        items:
          type: string
        type: array
    required:
    - id
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@higgstv.com
    name: API Support
    url: http://www.higgstv.com
  description: Session-based authentication
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: HiggsTV API Server
  version: "1.0"
paths:
  /apis/addchannel:
    post:
      consumes:
      - application/json
      description: 建立新頻道（需要登入）
      parameters:
      - description: 新增頻道請求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.AddChannelRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 需要登入" example({"state":1,"code":1})
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiAuth: []
      summary: 新增頻道
      tags:
      - 頻道
...
  /apis/getchannel/{id}:
    get:
      description: 根據頻道 ID 取得頻道詳細資訊
      parameters:
      - description: 頻道 ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 頻道不存在" example({"state":1,"code":2})
          schema:
            additionalProperties: true
            type: object
      summary: 取得單一頻道
      tags:
      - 頻道
  /apis/getchannelinfo/{id}:
    get:
      description: 取得頻道詳細資訊，包含擁有者基本資訊
      parameters:
      - description: 頻道 ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 頻道不存在" example({"state":1,"code":2})
          schema:
            additionalProperties: true
            type: object
      summary: 取得頻道資訊（含擁有者）
      tags:
      - 頻道
  /apis/getchannels:
    get:
      description: 取得公開頻道列表，支援過濾和排序
      parameters:
      - description: 只列出特定使用者的頻道（username）
        in: query
        name: user
        type: string
      - description: 以名稱模糊搜尋
        in: query
        name: q
        type: string
      - description: 是否只顯示有節目的頻道（0/1）
        in: query
        name: has_contents
        type: string
      - collectionFormat: csv
        description: 要排除的頻道類型陣列
        in: query
        items:
          type: string
        name: ignore_types
        type: array
      - description: 排序欄位（例如：last_modified, name）
        in: query
        name: sort
        type: string
      - description: 是否遞減排序（0/1）
        in: query
        name: desc
        type: string
      - description: 分頁起始 index
        in: query
        name: start
        type: integer
      - description: 限制筆數
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功回應
          schema:
            additionalProperties: true
            type: object
      summary: 取得頻道列表
      tags:
      - 頻道
  /apis/getownchannels:
    get:
      description: 取得當前登入使用者擁有的所有頻道（需要登入），支援 q 和 types[] 過濾
      parameters:
      - description: 關鍵字搜尋（頻道名稱）
        in: query
        name: q
        type: string
      - collectionFormat: csv
        description: 頻道類型陣列（例如：default,unclassified）
        in: query
        items:
          type: string
        name: types
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: 需要登入" example({"state":1,"code":1})
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiAuth: []
      summary: 取得自己的頻道列表
      tags:
      - 頻道
...
  /apis/savechannel:
    post:
      consumes:
      - application/json
      description: 更新頻道名稱、描述和標籤（需登入，需有權限）
      parameters:
      - description: 儲存頻道請求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.SaveChannelRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 權限不足" example({"state":1,"code":2})
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiAuth: []
      summary: 儲存頻道
      tags:
      - 頻道
...
  /apis/setchannelowner:
    post:
      consumes:
      - application/json
      description: 新增共用者，呼叫者需為 admin（優先使用 email 參數）
      parameters:
      - description: 設定頻道擁有者請求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.SetChannelOwnerRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 權限不足" example({"state":1,"code":2})
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiAuth: []
      summary: 設定頻道擁有者
      tags:
      - 頻道
...
swagger: "2.0"